Modify method signature for showFlightsFrom()
Previous version of the method
(bc1f7ab182aa63230738901c99f1d1524ad250ec)
used an integer representation of the airport ID property as an argument.

This version modifies the signature, along with method's annotations, to
take the preconstructed Airport object as the argument.

This is achieved by modifying the route parameter name to match the name
of the object property, allowing Doctrine to automagically instantiate
the required object.