Implement partial filter funtionality
Implemented filtering based on departure and destination airports,
but for now ignoring the date of departure.

Further info:
While modeling the Flight entity, I've made a rash decision to store
date and time of flight departure / arrival as a single datetime entry.

When users want to find flights between two airports, they would
presumably wish to do so based on a date, and not on a time frame.
Presumably.

This leads to the problem of finding a way to filter through the
datetime entries based only on provided date info, but seems more like
a task of finding a proper way of filtering on the database side and
maybe finding a way to do so with doctrine, rather than perhaps an
actual issue/bug.

If this can't be resolved in a manner similar to ones described here,
alternative would be to split out the datetime fields in database into
separate fields for date and time, and make appropriate changes in
Flight entity.