Create a feature for adding a new airport
This test will succeed in adding a new airport, but will eventually
fail in the end, because it expects the newly created airport to have
ID of 1 (actually, the test expects to end up on page representing the
airport with ID of 1).
Fail will occur if there is a flight record in db already.