Refactor sexyairlines to use Country entity
Instead of using a string representation of Country's ID property and
using that to retrieve the Country entity all over the place, we're
injectin the entity itself.

This should clean up the code a bit, because sexyairlines shouldn't need
to worry about constructing the country when it needs it, but rather
just receive the object it needs.

Service configuration file is being updated to reflect this change.